refactor: move updatePnpmCatalogDependency to its own module
